Material Cost Util Crawler
Figuring out what to organize within the this project and starting out with this package. Going through and changing some of the code into xpath in order to reduce how much many lines there is and simplify it a bit to make it easier to read. Going through and adding comments into my code as well as reducing how many lines in a function.